+++ src/plutarch-onchain-lib/lib/Plutarch/Core/Crypto.hs
@@ -17,17 +17,10 @@ import           Data.ByteString         (ByteString)
 import qualified Data.ByteString         as BS
 import           Data.ByteString.Short   (fromShort)
 import           Data.Word               (Word8)
-import           Plutarch.Builtin.Bool           (pif)
-import Plutarch.Internal.Eq ((#==))
-import           Plutarch.Builtin.ByteString     (PByteString, plengthBS,
-                                                 psliceBS)
 import           Plutarch.Builtin.Crypto (pblake2b_224, pkeccak_256)
-import           Plutarch.Builtin.Integer        (PInteger)
-import Plutarch.Internal.Lift (pconstant)
 import           Plutarch.Script         (Script (unScript))
 import qualified PlutusCore.Crypto.Hash  as Hash
 import           PlutusLedgerApi.Common  (serialiseUPLC)
-import Plutarch.Internal.Numeric (pmod)
 import Plutarch.Internal.Term (punsafeBuiltin)
 import Plutarch.Prelude
 import PlutusCore qualified as PLC
