+++ src/plutarch-onchain-lib/lib/Plutarch/Core/Value.hs
@@ -1,0 +1,135 @@
+{-# LANGUAGE CPP                  #-}
+{-# LANGUAGE PolyKinds            #-}
+{-# LANGUAGE QualifiedDo          #-}
+{-# LANGUAGE UndecidableInstances #-}
+{-# LANGUAGE DataKinds            #-}
+{-# LANGUAGE TypeFamilies         #-}
+{-# LANGUAGE TypeOperators        #-}
+{-# LANGUAGE ScopedTypeVariables  #-}
+{-# LANGUAGE FlexibleContexts     #-}
+{-# LANGUAGE GADTs                #-}
+{-# LANGUAGE RankNTypes           #-}
+
+module Plutarch.Core.Value where
+
+import Plutarch.Prelude
+import Plutarch.LedgerApi.V3
+import Plutarch.Core.PByteString(pisPrefixOf)
+import Plutarch.LedgerApi.Value
+import qualified Plutarch.LedgerApi.Value as Value
+import qualified Plutarch.LedgerApi.AssocMap as AssocMap
+import Plutarch.Internal.Term (punsafeCoerce)
+import Plutarch.Core.Internal.Builtins ( pmapData, ppairDataBuiltinRaw )
+
+{- | Finds the associated Currency symbols that contain token names prefixed with the ByteString.
+-}
+pfindCurrencySymbolsByTokenPrefix ::
+  forall
+    (anyOrder :: KeyGuarantees)
+    (anyAmount :: AmountGuarantees).
+  ClosedTerm
+    ( PValue anyOrder anyAmount
+        :--> PByteString
+        :--> PBuiltinList (PAsData PCurrencySymbol)
+    )
+pfindCurrencySymbolsByTokenPrefix = phoistAcyclic $
+  plam $ \value prefix ->
+    plet (pisPrefixOf # prefix) $ \prefixCheck ->
+      let mapVal = pto (pto value)
+          isPrefixed = pfilter # plam (\csPair ->
+            pany # plam (\tkPair ->
+              pmatch (pto (pfromData $ pfstBuiltin # tkPair)) $ \(PDataNewtype tkn) ->
+                  prefixCheck # pfromData tkn
+              ) # pto (pfromData (psndBuiltin # csPair))
+            ) # mapVal
+       in pmap # pfstBuiltin # isPrefixed
+
+{- | Finds the associated Currency symbols that contain the given token
+  name.
+-}
+pfindCurrencySymbolsByTokenName ::
+  forall
+    (anyOrder :: KeyGuarantees)
+    (anyAmount :: AmountGuarantees).
+  ClosedTerm
+    ( PValue anyOrder anyAmount
+        :--> PTokenName
+        :--> PBuiltinList (PAsData PCurrencySymbol)
+    )
+pfindCurrencySymbolsByTokenName = phoistAcyclic $
+  plam $ \value tn ->
+      let mapVal = pto (pto value)
+          hasTn = pfilter # plam (\csPair -> pany # plam (\tk -> tn #== pfromData (pfstBuiltin # tk)) # pto (pfromData (psndBuiltin # csPair))) # mapVal
+       in pmap # pfstBuiltin # hasTn
+
+-- | Checks if a Currency Symbol is held within a Value
+phasDataCS ::
+  forall
+    (anyOrder :: KeyGuarantees)
+    (anyAmount :: AmountGuarantees).
+  ClosedTerm
+    (PAsData PCurrencySymbol :--> PValue anyOrder anyAmount :--> PBool)
+phasDataCS = phoistAcyclic $
+  plam $ \symbol value ->
+    pany # plam (\tkPair -> (pfstBuiltin # tkPair) #== symbol) #$ pto (pto value)
+
+phasCS ::
+  forall
+    (anyOrder :: KeyGuarantees)
+    (anyAmount :: AmountGuarantees).
+  ClosedTerm
+    (PValue anyOrder anyAmount :--> PCurrencySymbol :--> PBool)
+phasCS = phoistAcyclic $
+  plam $ \value symbol ->
+    pany # plam (\tkPair -> pfromData (pfstBuiltin # tkPair) #== symbol) #$ pto (pto value)
+
+-- | Checks that a Value contains all the given CurrencySymbols.
+pcontainsCurrencySymbols ::
+  forall
+    (anyOrder :: KeyGuarantees)
+    (anyAmount :: AmountGuarantees).
+  ClosedTerm
+    ( PValue anyOrder anyAmount
+        :--> PBuiltinList (PAsData PCurrencySymbol)
+        :--> PBool
+    )
+pcontainsCurrencySymbols = phoistAcyclic $
+  plam $ \inValue symbols ->
+    let value = pmap # pfstBuiltin #$ pto $ pto inValue
+        containsCS = plam $ \cs -> pelem # cs # value
+     in pall # containsCS # symbols
+
+-- | Probably more effective than `plength . pflattenValue`
+pcountOfUniqueTokens ::
+  forall
+    (keys :: KeyGuarantees)
+    (amounts :: AmountGuarantees)
+    (s :: S).
+  Term s (PValue keys amounts :--> PInteger)
+pcountOfUniqueTokens = phoistAcyclic $
+  plam $ \val ->
+    let tokensLength = plam (\pair -> pmatch (pfromData $ psndBuiltin # pair) $ \(PMap tokens) -> plength # tokens)
+     in pmatch val $ \(PValue val') ->
+          pmatch val' $ \(PMap csPairs) -> pfoldl # plam (\acc x -> acc + (tokensLength # x)) # 0 # csPairs
+
+-- | Subtracts one Value from another
+psubtractValue ::
+  forall
+    (amounts :: AmountGuarantees)
+    (s :: S).
+  Term s (PValue 'Sorted amounts) ->
+  Term s (PValue 'Sorted amounts) ->
+  Term s (PValue 'Sorted 'NonZero)
+a `psubtractValue` b = pnormalize #$ Value.punionResolvingCollisionsWith AssocMap.NonCommutative # plam (-) # a # b
+
+-- | Constructs a singleton `PValue` with the given currency symbol, token name, and amount.
+--
+-- @param currencySymbol The currency symbol of the token.
+-- @param tokenName The name of the token.
+-- @param amount The amount of the token.
+--
+-- @return A singleton `PValue` containing the specified currency symbol, token name, and amount.
+pvalueSingleton :: Term s (PAsData PCurrencySymbol) -> Term s (PAsData PTokenName) -> Term s (PAsData PInteger) -> Term s (PAsData (PValue 'Sorted 'Positive))
+pvalueSingleton currencySymbol tokenName amount =
+  let innerValue = pcons @PBuiltinList # (ppairDataBuiltin # tokenName # amount) # pnil
+  in punsafeCoerce $ pmapData # (pcons @PBuiltinList # (ppairDataBuiltinRaw # pforgetData currencySymbol #$ pmapData # punsafeCoerce innerValue) # pnil)
